fix: common trait to read request headers
