+++ src/Http/Controller/AbstractAdminAjaxCorsController.php
@@ -30,10 +30,11 @@ use PrestaShop\Module\PsAccounts\Polyfill\Traits\AdminController\IsAnonymousAllo
 use PrestaShop\Module\PsAccounts\Polyfill\Traits\Controller\AjaxRender;
 use PrestaShop\Module\PsAccounts\Service\AdminTokenService;
 
-abstract class AbstractAdminAjaxController extends ModuleAdminController
+abstract class AbstractAdminAjaxCorsController extends ModuleAdminController
 {
     use AjaxRender;
     use IsAnonymousAllowed;
+    use GetHeader;
 
     const HEADER_AUTHORIZATION = 'X-PrestaShop-Authorization';
 
@@ -142,46 +143,6 @@ abstract class AbstractAdminAjaxController extends ModuleAdminController
             Logger::getInstance()->error($e);
             throw new UnauthorizedException($errorMsg);
         }
-    }
-
-    /**
-     * @param string $header
-     *
-     * @return string|null
-     */
-    protected function getRequestHeader($header)
-    {
-        $headerValue = null;
-
-        $headerKey = 'HTTP_' . strtoupper(str_replace('-', '_', $header));
-
-        if (array_key_exists($headerKey, $_SERVER)) {
-            $headerValue = $_SERVER[$headerKey];
-        }
-
-        if (null === $headerValue) {
-            $headerValue = $this->getApacheHeader($header);
-        }
-
-        return $headerValue;
-    }
-
-    /**
-     * @param string $header
-     *
-     * @return string|null
-     */
-    protected function getApacheHeader($header)
-    {
-        if (function_exists('apache_request_headers')) {
-            $headers = getallheaders();
-            //$header = preg_replace('/PrestaShop/', 'Prestashop', $header);
-            if (array_key_exists($header, $headers)) {
-                return $headers[$header];
-            }
-        }
-
-        return null;
     }
 
     /**
